backend/requirements.txt
fastapi
uvicorn[standard]

# Database
asyncpg
databases
sqlalchemy
psycopg2-binary

# Utils
python-dotenv
pyjwt
pyyaml
aiofiles

# LLM
langchain
langgraph
langchain-community
langchain-mistralai
langchain_groq
